Move tests in test directory + migrate on pytest
