+++ .github/workflows/mindsdb.yml
@@ -77,10 +77,10 @@ jobs:
           echo -e "\n===============\nDB HANDLERS\n===============\n"
 
           echo -e "\n===============MySQL Handler===============\n"
-          python mindsdb/integrations/handlers/mysql_handler/tests/test_mysql_handler.py
+          pytest -svx tests/handler_tests/test_mysql_handler.py
 
           echo -e "\n===============PostgreSQL Handler===============\n"
-          python mindsdb/integrations/handlers/postgres_handler/tests/test_postgres_handler.py
+          pytest -svx tests/handler_tests/test_postres_handler.py
 
           mkdir -p ~/.ssh/
           echo "$DB_MACHINE_KEY" > ~/.ssh/db_machine
