+++ docker/db_images/README.md
@@ -1,0 +1,1 @@
+### This directory contains all data to build docker images for db handlers testing
