+++ docker/db_images/mysql/sql-scripts/prepare.sql
