+++ docker/db_images/postgres/sql-scripts/prepare.sql
